chore: thêm eslint và prettier
