+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+    "plugins": [
+        "prettier-plugin-tailwindcss"
+    ]
+}
